requirements.txt
pillow==11.3.0
numpy==2.0.2
opencv-python==4.12.0.88
PyAutoGUI==0.9.54
pynput==1.8.1
mss==10.1.0